+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `Spot` module
+- Add `Http.QueryParameters` module
 
 ## 8.0.0 - 2023-08-11
 - [**BC**] Use net 7.0